moved things around + fixed build errors
