+++ Assets/Textures/Items/CosmostoneShowers/SparklingBomb.png
