+++ Assets/Textures/Items/CosmostoneShowers/SparklingDynamite.png
