+++ Assets/Textures/Items/CosmostoneShowers/SparklingGrenade.png
