+++ Assets/Textures/Items/CosmostoneShowers/SparklingGrenadeProjectile_Glow.png
