+++ Assets/Textures/NPCs/CosmostoneShowers/Angelatin.png
