+++ Content/Items/CosmostoneShowers/Stargel.cs
@@ -1,8 +1,13 @@
-﻿namespace TwilightEgress.Content.Items.Materials
+﻿using Terraria;
+using Terraria.ModLoader;
+
+namespace TwilightEgress.Content.Items.CosmostoneShowers
 {
     public class Stargel : ModItem
     {
         public new string LocalizationCategory => "Items.Materials";
+
+        public override string Texture => base.Texture.Replace("Content", "Assets/Textures");
 
         public override void SetStaticDefaults() => Item.ResearchUnlockCount = 5;
