--- Content/Items/Weapons/Ranged/SparklingExplosives/SparklingBomb.cs
@@ -1,46 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TwilightEgress.Content.Items.Materials;
-
-namespace TwilightEgress.Content.Items.Weapons.Ranged.SparklingExplosives
-{
-    public class SparklingBomb : ModItem
-    {
-        public new string LocalizationCategory => "Items.Weapons.Ranged";
-
-        public override void SetStaticDefaults()
-        {
-            ItemID.Sets.ItemsThatCountAsBombsForDemolitionistToSpawn[Type] = true;
-            Item.ResearchUnlockCount = 99;
-        }
-
-        public override void SetDefaults()
-        {
-            Item.useStyle = ItemUseStyleID.Swing;
-            Item.shootSpeed = 12f;
-            Item.shoot = ProjectileID.Bomb;
-            Item.width = 22;
-            Item.height = 32;
-            Item.maxStack = Item.CommonMaxStack;
-            Item.consumable = true;
-            Item.UseSound = SoundID.Item1;
-            Item.useAnimation = 40;
-            Item.useTime = 40;
-            Item.noUseGraphic = true;
-            Item.noMelee = true;
-            Item.value = Item.sellPrice(silver: 1);
-            Item.rare = ItemRarityID.White;
-        }
-
-        public override void AddRecipes()
-        {
-            CreateRecipe()
-                .AddIngredient<Stargel>()
-                .AddIngredient(ItemID.Bomb)
-                .Register();
-        }
-    }
-}
